requirements.txt
pandas
numpy
requests
tqdm
jaro-winkler